feat(fav2): moved logic to create icon stack to action
action is used in v1 and v2 endpoints for font-awesome v5 icons
